reorganise files and folders
