+++ src/js/commonFunctions/errorHandling.mjs
