+++ src/js/components/postClass.mjs
@@ -1,7 +1,7 @@
 import parser from "../tools/parser.mjs";
 import formatDate from "../tools/dateStyler.mjs";
-import { accessToken } from "./storedKeys.mjs";
-import { viewSpecificPost } from "../auth/commonFunctions/api.mjs";
+import { accessToken } from "../constants/storedKeys.mjs";
+import { viewSpecificPost } from "../commonFunctions/api.mjs";
 
 export class Post {
   constructor(
