+++ src/js/constants/apiUrl.mjs
@@ -1,0 +1,1 @@
+export const BASE_URL = "https://nf-api.onrender.com";
