+++ src/js/homepage/homepageScript.js
@@ -1,6 +1,5 @@
-import { fetchPosts } from "../auth/commonFunctions/api.mjs";
-import { createPost } from "../auth/commonFunctions/api.mjs";
-import { accessToken } from "../components/storedKeys.mjs";
+import { fetchPosts, createPost } from "../commonFunctions/api.mjs";
+import { accessToken } from "../constants/storedKeys.mjs";
 import { Post } from "../components/postClass.mjs";
 
 const listOfPosts = await fetchPosts(accessToken);
